requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-appconf==1.0.6
django-compressor==4.4
django-favicon==0.1.4
django-import-export==3.3.5
django-uuidfield==0.5.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
psycopg2==2.9.9
PyYAML==6.0.1
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
